Use alpine build image for easier platform support
The existing docker image cannot easily be built for multi-arch using
`docker buildx` or similar since it encodes the Rust target architecture
using the `ARCH` build-arg. `docker buildx` provides the target
architecture using the `TARGETARCH` build-arg but uses `amd64` and
`arm64` compared to Rust using `x86_64` and `aarch64`.
To make this easier I've changed the builder image to the Alpine based
version so that `--target` is not required to produce a static
musl-based binary (since it's the default on Alpine).